handwriting_requirements.txt
streamlit
streamlit-drawable-canvas
pandas
google-generativeai
pillow